Fixed issue #4 in an attempt to improve receipts

+++ src/main/java/wtf/meier/tariff/interpreter/model/Price.kt
@@ -1,5 +1,5 @@
 package wtf.meier.tariff.interpreter.model
 
 data class Price(
-    val credit: Long,
+    val credit: Int,
 )
